Move action tests to their own folder
